Updates for smallStatements
Separate statement and smallStatements into two events.  ability to parse recognise a smallStatement request and the extract maxSmallQuerySize from it. Implementation in place for sending all results of a small statement in one smallResultsSet event. test added (005) to check this behaviour. Currnetly this is not workign as the sql and parameters are at the wrogn level for reading them from the payload for smallStatements. and threfore is a todo to stop a runnign query and ignore further results if there are more than MaxSizeQuerySize